Acoustic: Rename `fwi_operators.py` -> `acoustic/operators.py`
