+++ examples/acoustic/__init__.py
@@ -1,1 +1,2 @@
+from .operators import *  # noqa
 from .wavesolver import *  # noqa
