+++ examples/acoustic/operators.py
